requirements.txt
Click==7.0
Flask==1.1.1
Flask-Limiter==1.1.0
flask-talisman==0.7.0
itsdangerous==1.1.0
Jinja2==2.10.3
limits==1.3
MarkupSafe==1.1.1
pymongo==3.9.0
python-dotenv==0.10.3
six==1.13.0
Werkzeug==0.16.0
yapf==0.28.0
